integer

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -4,7 +4,7 @@
  * @n: value integer
  * Return
  */
-void print_number(int n)
+void print_number(long int n)
 {
 	if (n < 0)
 	{
